third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-cancel-hover-state-expected.txt

Gestures go here
Tests that gesture tapcancel clears hover properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


See http://crbug.com/316974 for details

Verify hover, active aren't initially set
PASS isBoxDefault() is true
tapcancel on element when it is Active should keep hover
PASS isBoxHoverActive() is true
PASS isBoxHovered() is true
tapcancel outside element when it is Active should clear it
PASS isBoxHoverActive() is true
PASS isBoxDefault() is true
tapcancel on element when it is hovered but not active should keep hover
PASS isBoxHovered() is true
PASS isBoxHovered() is true
tapcancel outside element when it is hovered but not active should keep hover
PASS isBoxHovered() is true
PASS isBoxHovered() is true
PASS successfullyParsed is true

TEST COMPLETE
